Handcraft Sustainable Clothing
At CHARKHTALES, we are driven by the spirit to bring the emporium of fashion, which combines ethical craftsmanship and time-honored designs to encapsulate timeless fashion, delivered right to your doorstep.

We aim to make Khadi, the revolutionary fabric, an integral part of the fashion industry, and deliver good quality products at competitive prices without compromising on our social and ethical responsibilities. With our extensive product portfolio, hassle-free customer service, and doorstep delivery, the whole team of CHARKHTALES is dedicated to reviving the ancient process of handspun fabric and making it popular on a global platform. Let’s Embrace Khadi!

CHARKHTALES supports sustainable practices and aims at nurturing the beauty of traditional craftsmanship. We emphasize curating merchandise from the best Khadi brands across the globe with an intent to deliver products that are unique, comfortable, and of heirloom quality.

Explore our range of sustainable clothing, couture, and designer wear today!